# contracts/openapi-v1.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LinkResponse": {
        "properties": {
          "link_id": {
            "title": "Link Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "point_id": {
            "title": "Point Id",
            "type": "string"
          },
          "reference_id": {
            "title": "Reference Id",
            "type": "string"
          },
          "relation": {
            "title": "Relation",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          }
        },
        "required": [
          "link_id",
          "reference_id",
          "organization_id",
          "site_id",
          "point_id",
          "relation"
        ],
        "title": "LinkResponse",
        "type": "object"
      },
      "LinkUpsertRequest": {
        "properties": {
          "organization_id": {
            "minLength": 1,
            "title": "Organization Id",
            "type": "string"
          },
          "point_id": {
            "minLength": 1,
            "title": "Point Id",
            "type": "string"
          },
          "relation": {
            "default": "observes",
            "minLength": 1,
            "title": "Relation",
            "type": "string"
          },
          "site_id": {
            "minLength": 1,
            "title": "Site Id",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "site_id",
          "point_id"
        ],
        "title": "LinkUpsertRequest",
        "type": "object"
      },
      "MappingItemRequest": {
        "properties": {
          "mapping_id": {
            "minLength": 1,
            "title": "Mapping Id",
            "type": "string"
          },
          "source_key": {
            "minLength": 1,
            "title": "Source Key",
            "type": "string"
          },
          "target_point_id": {
            "minLength": 1,
            "title": "Target Point Id",
            "type": "string"
          },
          "transform": {
            "default": "identity",
            "minLength": 1,
            "title": "Transform",
            "type": "string"
          }
        },
        "required": [
          "mapping_id",
          "source_key",
          "target_point_id"
        ],
        "title": "MappingItemRequest",
        "type": "object"
      },
      "MappingReplaceRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MappingItemRequest"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "organization_id": {
            "minLength": 1,
            "title": "Organization Id",
            "type": "string"
          },
          "site_id": {
            "minLength": 1,
            "title": "Site Id",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "site_id",
          "items"
        ],
        "title": "MappingReplaceRequest",
        "type": "object"
      },
      "MappingResponse": {
        "properties": {
          "mapping_id": {
            "title": "Mapping Id",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "reference_id": {
            "title": "Reference Id",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "target_point_id": {
            "title": "Target Point Id",
            "type": "string"
          },
          "transform": {
            "title": "Transform",
            "type": "string"
          }
        },
        "required": [
          "mapping_id",
          "reference_id",
          "organization_id",
          "site_id",
          "source_key",
          "target_point_id",
          "transform"
        ],
        "title": "MappingResponse",
        "type": "object"
      },
      "ReferenceResponse": {
        "properties": {
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "reference_id": {
            "title": "Reference Id",
            "type": "string"
          },
          "site_id": {
            "title": "Site Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "reference_id",
          "organization_id",
          "site_id",
          "device_id",
          "label",
          "metadata",
          "updated_at"
        ],
        "title": "ReferenceResponse",
        "type": "object"
      },
      "ReferenceUpsertRequest": {
        "properties": {
          "device_id": {
            "minLength": 1,
            "title": "Device Id",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "organization_id": {
            "minLength": 1,
            "title": "Organization Id",
            "type": "string"
          },
          "site_id": {
            "minLength": 1,
            "title": "Site Id",
            "type": "string"
          }
        },
        "required": [
          "organization_id",
          "site_id",
          "device_id",
          "label"
        ],
        "title": "ReferenceUpsertRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Wave 2 baseline for device references, mappings, and links. Path versioned endpoints under /api/v1.",
    "title": "Reference API Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/device-references": {
      "get": {
        "operationId": "list_references_api_v1_device_references_get",
        "parameters": [
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReferenceResponse"
                  },
                  "title": "Response List References Api V1 Device References Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List References"
      }
    },
    "/api/v1/device-references/{reference_id}": {
      "get": {
        "operationId": "get_reference_api_v1_device_references__reference_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Reference"
      },
      "put": {
        "operationId": "upsert_reference_api_v1_device_references__reference_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Reference"
      }
    },
    "/api/v1/device-references/{reference_id}/links": {
      "get": {
        "operationId": "list_links_api_v1_device_references__reference_id__links_get",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LinkResponse"
                  },
                  "title": "Response List Links Api V1 Device References  Reference Id  Links Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Links"
      }
    },
    "/api/v1/device-references/{reference_id}/links/{link_id}": {
      "put": {
        "operationId": "upsert_link_api_v1_device_references__reference_id__links__link_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LinkUpsertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Link"
      }
    },
    "/api/v1/device-references/{reference_id}/mappings": {
      "get": {
        "operationId": "list_mappings_api_v1_device_references__reference_id__mappings_get",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Organization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "site_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Site Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MappingResponse"
                  },
                  "title": "Response List Mappings Api V1 Device References  Reference Id  Mappings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Mappings"
      },
      "put": {
        "operationId": "replace_mappings_api_v1_device_references__reference_id__mappings_put",
        "parameters": [
          {
            "in": "path",
            "name": "reference_id",
            "required": true,
            "schema": {
              "title": "Reference Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MappingReplaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MappingResponse"
                  },
                  "title": "Response Replace Mappings Api V1 Device References  Reference Id  Mappings Put",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Mappings"
      }
    },
    "/healthz": {
      "get": {
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    }
  }
}